Update main.py
Only outputs edge maps (the png files) in a single directory, added support for imagenet datasets using this tool - https://github.com/mf1024/ImageNet-Datasets-Downloader. Essentially now, you point it to the directory of your imagenet dataset and it spits out all the edge maps in a single folder. The edge maps have the same names as the imagenet images but are png files. Also included a bit of code to delete .ipynb checkpoints from the affected directories - this was to avoid errors generating edge maps when running in a notebook environment.